+++ djangojs/utils.py
@@ -204,13 +204,18 @@ class ContextSerializer(object):
             'permissions': tuple(),
         }
         if 'django.contrib.sessions.middleware.SessionMiddleware' in settings.MIDDLEWARE_CLASSES:
-            data['user'].update({
-                'username': request.user.username,
-                'is_authenticated': request.user.is_authenticated(),
-                'is_staff': request.user.is_staff,
-                'is_superuser': request.user.is_superuser,
-                'permissions': tuple(request.user.get_all_permissions())
-            })
+            user = request.user
+            data['user']['is_authenticated'] = user.is_authenticated()
+            if hasattr(user, 'username'):
+                data['user']['username'] = user.username
+            elif hasattr(user, 'get_username'):
+                data['user']['username'] = user.get_username()
+            if hasattr(user, 'is_staff'):
+                data['user']['is_staff'] = user.is_staff
+            if hasattr(user, 'is_superuser'):
+                data['user']['is_superuser'] = user.is_superuser
+            if hasattr(user, 'get_all_permissions'):
+                data['user']['permissions'] = tuple(user.get_all_permissions())
 
 
 class StorageGlobber(object):